feat(cherry-blossom-qrcode): implement Cherry Blossom QR code animation with WebGPU support
