+++ src/animations/cherry-blossom-qrcode/constants.ts
@@ -1,0 +1,49 @@
+// Background and container colors
+export const COLORS = {
+  background: '#f7f7f7',
+} as const;
+
+export const CONTAINER_BG = COLORS.background;
+export const DEFAULT_QR_CONTENT = 'https://enzo.fyi';
+
+// Color palette for lighting
+export const PALETTE = {
+  skyZenith: { r: 0.82, g: 0.88, b: 0.92 },
+  skyHorizon: { r: 0.91, g: 0.93, b: 0.91 },
+  sun: { r: 1.15, g: 1.05, b: 0.95 },
+  skyFill: { r: 0.85, g: 0.9, b: 0.95 },
+  bounce: { r: 0.5, g: 0.65, b: 0.42 },
+} as const;
+
+// Block/cube dimensions
+export const BLOCK_SIZE = 0.0245;
+export const CUBE_HEIGHT = BLOCK_SIZE;
+
+// Tree structure parameters
+export const TRUNK_RADIUS = 2.5;
+export const TRUNK_LAYERS = 12;
+export const MAX_CANOPY_LAYERS = 12;
+export const CANOPY_OUTER_RADIUS_FACTOR = 0.46;
+
+// Grid limits
+export const MAX_GRID_SIZE = 41;
+export const MAX_BLOCKS = MAX_GRID_SIZE * MAX_GRID_SIZE * 18;
+
+// Camera angles for 3D isometric view
+export const ISO_ANGLE_Y = 0.78;
+export const ISO_ANGLE_X = -0.55;
+
+// Camera angles for 2D flat view (top-down for QR scanning)
+export const FLAT_ANGLE_Y = 0.0;
+export const FLAT_ANGLE_X = -1.5708; // -π/2
+
+// Animation
+export const LERP_SPEED = 4.0;
+
+// View scaling
+export const VIEW_SCALE_3D = 1.6;
+export const VIEW_SCALE_2D = 2.1;
+
+// Centering offsets for 2D view
+export const Y_OFFSET_2D = 0.08;
+export const X_OFFSET_2D = 0.015;
