+++ src/animations/cherry-blossom-qrcode/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useWebGPU } from './use-web-gpu';
