+++ src/animations/cherry-blossom-qrcode/shaders/index.ts
@@ -1,0 +1,4 @@
+export { blocksVertexShader } from './blocks-vertex';
+export { blocksFragmentShader } from './blocks-fragment';
+export { skyVertexShader, skyFragmentShader } from './sky';
+export { shadowVertexShader, shadowFragmentShader } from './shadow';
