+++ src/animations/cherry-blossom-qrcode/shaders/sky.ts
@@ -1,0 +1,39 @@
+import { uniformsStruct } from './helpers';
+
+export const skyVertexShader = /* wgsl */ `
+${uniformsStruct}
+
+struct SkyOut {
+  @builtin(position) position: vec4f,
+  @location(0) uv: vec2f,
+}
+
+@group(0) @binding(0) var<uniform> uniforms: Uniforms;
+
+@vertex
+fn main(@builtin(vertex_index) vi: u32) -> SkyOut {
+  var tri = array<vec2f, 3>(
+    vec2f(-1.0, -1.0),
+    vec2f(3.0, -1.0),
+    vec2f(-1.0, 3.0)
+  );
+  let p = tri[vi];
+  var o: SkyOut;
+  o.position = vec4f(p, 1.0, 1.0);
+  o.uv = vec2f(p.x * 0.5 + 0.5, 0.5 - p.y * 0.5);
+  return o;
+}
+`;
+
+// Background color matching container (#f7f7f7)
+export const skyFragmentShader = /* wgsl */ `
+${uniformsStruct}
+
+@group(0) @binding(0) var<uniform> uniforms: Uniforms;
+
+@fragment
+fn main(@location(0) uv: vec2f) -> @location(0) vec4f {
+  // Match container background #f7f7f7 = rgb(247, 247, 247)
+  return vec4f(0.969, 0.969, 0.969, 1.0);
+}
+`;
