+++ src/animations/registry.ts
@@ -37,6 +37,7 @@ import { BottomBarSkia } from './bottom-bar-skia';
 import { CalendarDays } from './calendar-days';
 import { CardShaderReflections } from './card-shader-reflections';
 import { CheckboxInteractions } from './checkbox-interactions';
+import { CherryBlossomQRCode } from './cherry-blossom-qrcode';
 import { CircularCarousel } from './circular-carousel';
 import { ClerkToast } from './clerk-toast';
 import { ClockTimePicker } from './clock-time-picker';
@@ -69,7 +70,6 @@ import { GeometryButton } from './geometry-button';
 import { GitHubContributions } from './github-contributions';
 import { GitHubOnboarding } from './github-onboarding';
 import { GitHubTerrain } from './github-terrain';
-import { CherryBlossomQRCode } from './cherry-blossom-qrcode';
 import { GLTransitions } from './gl-transitions';
 import { GridVisualizer } from './grid-visualizer';
 import { ImageCropper } from './image-cropper';
